requirements.txt
fastapi
uvicorn
aiogram==3.13.0
python-dotenv